requirements.txt
colorzero==1.1
gpiozero==1.5.1
IMAPClient==2.1.0
pkg-resources==0.0.0
RPi.GPIO==0.7.0
RPLCD==1.2.2
six==1.14.0
